refactor(android): extract GattCallbackEvent sealed interface to own file
AndroidGattBridge reduced from 315 to 253 lines (-20%), below the 300-line
threshold. The GattCallbackEvent sealed interface hierarchy (11 event types)
is extracted to GattCallbackEvent.kt as a pure data-types file with no
dependencies on AndroidGattBridge state.

No behavioral changes — all event types and call sites unchanged.
